usb: dwc3: Use async resume during DPM complete
The DWC3 core device's runtime resume API has a lot of initialization
routines, which could contribute to delays during the PM resume path.
The DWC3 core's complete() callback will execute a synchronous resume,
which will cause the runtime resume to run in the PM resume context.
Remove the register accesses to DWC3_GUCTL3_SPLITDISABLE as it is not
applicable to our platform.  This allows for the dwc3 core complete to be
able to utilize the async runtime resume path.